+++ PurrKatEngine/src/PurrKatEngine/Events/CursorPosEvent.h
@@ -10,16 +10,16 @@ namespace PurrKatEngine
     class CursorPosEvent : public CursorEvent
     {
     public:
-        CursorPosEvent(float mouseX, float mouseY): m_MouseX(mouseX), m_MouseY(mouseY) {}
+        CursorPosEvent(double mouseX, double mouseY): m_MouseX(mouseX), m_MouseY(mouseY) {}
 
-        float GetX() const { return m_MouseX; }
-        float GetY() const { return m_MouseY; }
+        double GetX() const { return m_MouseX; }
+        double GetY() const { return m_MouseY; }
         
         EVENT_TO_STRING(CursorPosEvent, "{}, {}", m_MouseX, m_MouseY)
 
         EVENT_CLASS_TYPE(MouseMoved)
 
     private:
-        float m_MouseX, m_MouseY;
+        double m_MouseX, m_MouseY;
     };
 }
